requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via starlette
certifi==2025.10.5
    # via -r requirements.in
cffi==2.0.0
    # via cryptography
click==8.3.0
    # via uvicorn
colorama==0.4.6
    # via click
cryptography==46.0.3
    # via pyjwt
dnspython==2.8.0
    # via pymongo
fastapi==0.118.2
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.10
    # via anyio
numpy==2.3.3
    # via pandas
pandas==2.3.3
    # via -r requirements.in
pycparser==2.23
    # via cffi
pydantic==2.12.0
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.1
    # via pydantic
pydantic-settings==2.11.0
    # via -r requirements.in
pyjwt[crypto]==2.10.1
    # via -r requirements.in
pymongo==4.15.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via pydantic-settings
pytz==2025.2
    # via pandas
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.48.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
uvicorn==0.37.0
    # via -r requirements.in
